ADF Fintech Interview
1. input1: Length of the array
    input2: An array of positive integers
    Output Specification: Return the second largest number or -1 accordingly.

2. Maximum Contiguous Sequences
Given an array of N integers, determine all contiguous subsequences of positive numbers. Then write a function to find the sum of elements each sub-sequence and output the maximum sum value.
Input Specifiction:
input1: An integer N denoting length of array input2: An integer array of length N(-10000<=A[i]<=10000)
Output Specification
Return the maximum sum of contiguous sub-sequence of numbers

3. The Card Game
One day, Fred and his N friends were playing a card game in which each player throws a card with a number written on it. The cards are such that a number X is written on front of the card, and the negative of that number is written on the back side of the card. This game has the following rules:-

Each of N players is asked to throw a card. After all the N cards are thrown, Fred has to flip one or more cards in consecutive order, only once. Your task is to help Fred flip the cards in such a way that the sum of the numbers, on front face of the cards, is the maximum.